Create Student OOP
Create a class called student with the following attributes:
->Name
->RollNo
->Year
->Branch
->List of marks
Make only getters for all the attributes. Provide a constructor to initialize everything else but not the list of marks. Provide a setter for marks, in which you provide an int that is added to the list. Provide a toString() method that prints the first four details and the percentage marks of the person up to two decimal places by rounding off.
Override the hashCode and equals methods and add some duplicate RollNo and create a HashSet of students and print its size after each addition.
Add some comparators and sort the list of students on the basis of
->Average marks
->Branch, and within same branch by RollNo